Redesign id assignment in annotation_content
Annotation bodies do not require an id field unless they are linking to
content that must be downloaded, such as a transcript file.
Additionally, there may be use cases where a user needs or wants to
provided their own ID to the annotations themselves. Providing id
attrs for annotation_content and removing the requirment for a URL
allows us to accomodate these use cases.